+++ lang/src/org/partiql/lang/planner/transforms/AstToLogicalVisitorTransform.kt
@@ -1,13 +1,14 @@
 package org.partiql.lang.planner.transforms
 
+import com.amazon.ionelement.api.ionSymbol
 import org.partiql.lang.domains.PartiqlAst
 import org.partiql.lang.domains.PartiqlAstToPartiqlLogicalVisitorTransform
 import org.partiql.lang.domains.PartiqlLogical
 import org.partiql.lang.errors.Problem
+import org.partiql.lang.errors.ProblemHandler
 import org.partiql.lang.eval.physical.sourceLocationMetaOrUnknown
 import org.partiql.lang.planner.PlanningProblemDetails
-import org.partiql.lang.planner.abortQueryPlanning
-import org.partiql.lang.planner.abortUnimplementedFeature
+import org.partiql.lang.planner.handleUnimplementedFeature
 
 /**
  * Transforms an instance of [PartiqlAst.Statement] to [PartiqlLogical.Statement].  This representation of the query
@@ -18,20 +19,22 @@ import org.partiql.lang.planner.abortUnimplementedFeature
  * This conversion (and the logical algebra) are early in their lifecycle and so only a limited subset of SFW queries
  * are transformable.  See `AstToLogicalVisitorTransformTests` to see which queries are transformable.
  */
-internal fun PartiqlAst.Statement.toLogicalPlan(): PartiqlLogical.Plan =
+internal fun PartiqlAst.Statement.toLogicalPlan(problemHandler: ProblemHandler): PartiqlLogical.Plan =
     PartiqlLogical.build {
         plan(
-            AstToLogicalVisitorTransform.transformStatement(this@toLogicalPlan),
+            AstToLogicalVisitorTransform(problemHandler).transformStatement(this@toLogicalPlan),
             version = PLAN_VERSION_NUMBER
         )
     }
 
-private object AstToLogicalVisitorTransform : PartiqlAstToPartiqlLogicalVisitorTransform() {
+private class AstToLogicalVisitorTransform(
+    val problemHandler: ProblemHandler
+) : PartiqlAstToPartiqlLogicalVisitorTransform() {
 
     override fun transformExprSelect(node: PartiqlAst.Expr.Select): PartiqlLogical.Expr {
         checkForUnsupportedSelectClauses(node)
 
-        var algebra: PartiqlLogical.Bexpr = FromSourceToBexpr.convert(node.from)
+        var algebra: PartiqlLogical.Bexpr = node.from.toBexpr(this, problemHandler)
 
         algebra = node.fromLet?.let { fromLet ->
             PartiqlLogical.build {
@@ -83,7 +86,10 @@ private object AstToLogicalVisitorTransform : PartiqlAstToPartiqlLogicalVisitorT
                         // [SelectStarVisitorTransform]. Therefore, there is no need to support `SELECT *` here.
                         errAstNotNormalized("Expected SELECT * to be removed")
 
-                    is PartiqlAst.Projection.ProjectPivot -> TODO("PIVOT ...")
+                    is PartiqlAst.Projection.ProjectPivot -> {
+                        problemHandler.handleUnimplementedFeature(node, "PIVOT")
+                        INVALID_EXPR
+                    }
                 },
                 algebra,
                 node.project.metas
@@ -104,9 +110,9 @@ private object AstToLogicalVisitorTransform : PartiqlAstToPartiqlLogicalVisitorT
      */
     private fun checkForUnsupportedSelectClauses(node: PartiqlAst.Expr.Select) {
         when {
-            node.group != null -> abortUnimplementedFeature(node, "GROUP BY")
-            node.order != null -> abortUnimplementedFeature(node, "ORDER BY")
-            node.having != null -> abortUnimplementedFeature(node, "HAVING")
+            node.group != null -> problemHandler.handleUnimplementedFeature(node.group, "GROUP BY")
+            node.order != null -> problemHandler.handleUnimplementedFeature(node.order, "ORDER BY")
+            node.having != null -> problemHandler.handleUnimplementedFeature(node.having, "HAVING")
         }
     }
 
@@ -120,18 +126,18 @@ private object AstToLogicalVisitorTransform : PartiqlAstToPartiqlLogicalVisitorT
         }
 
     override fun transformStatementDml(node: PartiqlAst.Statement.Dml): PartiqlLogical.Statement {
+        require(node.operations.ops.isNotEmpty())
+
         // `INSERT` and `DELETE` statements are all that's needed for the current effort--and it just so
         // happens that these never utilize more than one DML operation anyway.  We don't need to
         // support more than one DML operation until we start supporting UPDATE statements.
-        if (node.operations.ops.size != 1) {
-            abortUnimplementedFeature(node, "more than one DML operation")
+        if (node.operations.ops.size > 1) {
+            problemHandler.handleUnimplementedFeature(node, "more than one DML operation")
         }
-        val dmlOp = node.operations.ops.single()
 
-        // All
-        return when (dmlOp) {
+        return when (val dmlOp = node.operations.ops.first()) {
             is PartiqlAst.DmlOp.Insert -> {
-                node.from?.let { TODO("FROM / INSERT ") }
+                node.from?.let { problemHandler.handleUnimplementedFeature(dmlOp, "UPDATE / INSERT") }
                 // Check for and block `INSERT INTO <tbl> VALUES (...)`  This is *no* way to support this
                 // within PartiQL itself since this flavor requires schema which we do not yet have.
                 // We block this by identifying (bag (list ...) ...) nodes which  is how the parser represents the
@@ -139,9 +145,9 @@ private object AstToLogicalVisitorTransform : PartiqlAstToPartiqlLogicalVisitorT
                 // `<< [ ... ] ... >>` and `BAG(LIST(...), ...)` those get blocked too.  This is probably just as well.
                 if (dmlOp.values is PartiqlAst.Expr.Bag) {
                     dmlOp.values.values.firstOrNull { it is PartiqlAst.Expr.List }?.let {
-                        abortQueryPlanning(
+                        problemHandler.handleProblem(
                             Problem(
-                                it.metas.sourceLocationMetaOrUnknown,
+                                node.metas.sourceLocationMetaOrUnknown,
                                 PlanningProblemDetails.InsertValuesDisallowed
                             )
                         )
@@ -150,68 +156,101 @@ private object AstToLogicalVisitorTransform : PartiqlAstToPartiqlLogicalVisitorT
 
                 PartiqlLogical.build {
                     dml(
-                        target = dmlTarget(transformExpr(dmlOp.target)),
+                        target = transformExpr(dmlOp.target),
                         operation = dmlInsert(),
                         rows = transformExpr(dmlOp.values),
                         metas = node.metas
                     )
                 }
             }
-            // INSERT single row with VALUE is disallowed use INSERT INTO <target> << <value>, <value> >> instead.
+            // INSERT single row with VALUE is disallowed. (This variation of INSERT might be removed in a future
+            // release of PartiQL.)
             is PartiqlAst.DmlOp.InsertValue -> {
-                abortQueryPlanning(
+                problemHandler.handleProblem(
                     Problem(
                         node.metas.sourceLocationMetaOrUnknown,
                         PlanningProblemDetails.InsertValueDisallowed
                     )
                 )
+                INVALID_STATEMENT
             }
             is PartiqlAst.DmlOp.Delete -> {
-                val relation = node.from?.let { FromSourceToBexpr.convert(it) }
                 if (node.from == null) {
                     // unfortunately, the AST allows malformations such as this however the parser should
                     // never actually create an AST for a DELETE statement without a FROM clause.
                     error("Malformed AST: DELETE without FROM (this should never happen)")
                 } else {
-                    val scan = relation as? PartiqlLogical.Bexpr.Scan
-                        ?: abortQueryPlanning(
-                            Problem(
-                                node.metas.sourceLocationMetaOrUnknown,
-                                PlanningProblemDetails.InvalidDmlTarget
+                    when (node.from) {
+                        is PartiqlAst.FromSource.Scan -> {
+                            val rowsSource = node.from.toBexpr(this, problemHandler) as PartiqlLogical.Bexpr.Scan
+                            val predicate = node.where?.let { transformExpr(it) }
+                            val rows = if (predicate == null) {
+                                rowsSource
+                            } else {
+                                PartiqlLogical.build { filter(predicate, rowsSource) }
+                            }
+
+                            PartiqlLogical.build {
+                                dml(
+                                    target = transformExpr(node.from.expr),
+                                    operation = dmlDelete(),
+                                    // This query returns entire rows which are to be deleted, which is unfortunate
+                                    // unavoidable without knowledge of schema. PartiQL embedders may apply a
+                                    // pass over the resolved logical (or later) plan that changes this to only
+                                    // include the primary keys of the rows to be deleted.
+                                    rows = bindingsToValues(
+                                        exp = id(rowsSource.asDecl.name.text, caseSensitive(), unqualified()),
+                                        query = rows,
+                                    ),
+                                    metas = node.metas
+                                )
+                            }
+                        }
+                        else -> {
+                            problemHandler.handleProblem(
+                                Problem(
+                                    node.from.metas.sourceLocationMetaOrUnknown,
+                                    PlanningProblemDetails.InvalidDmlTarget
+                                )
                             )
-                        )
-
-                    val predicate = node.where?.let { transformExpr(it) }
-                    val rows = if (predicate == null) scan else PartiqlLogical.build { filter(predicate, scan) }
-
-                    PartiqlLogical.build {
-                        dml(
-                            target = dmlTarget(scan.expr),
-                            operation = dmlDelete(),
-                            rows = bindingsToValues(
-                                // This returns entire rows which are to be deleted, which is very naive but
-                                // is unavoidable without knowledge of schema. PartiQL embedders may apply a
-                                // pass over the resolved logical (oor later) plan that changes this to only include
-                                // the primary keys of the rows to be deleted.
-                                exp = id(scan.asDecl.name.text, caseSensitive(), unqualified()),
-                                query = rows,
-                            ),
-                            metas = node.metas
-                        )
+                            INVALID_STATEMENT
+                        }
                     }
                 }
             }
             is PartiqlAst.DmlOp.Remove -> {
-                abortQueryPlanning(
-                    Problem(node.metas.sourceLocationMetaOrUnknown, PlanningProblemDetails.InvalidUseOfRemove)
+                problemHandler.handleProblem(
+                    Problem(dmlOp.metas.sourceLocationMetaOrUnknown, PlanningProblemDetails.UnimplementedFeature("REMOVE"))
                 )
+                INVALID_STATEMENT
             }
             is PartiqlAst.DmlOp.Set -> {
-                abortQueryPlanning(
-                    Problem(node.metas.sourceLocationMetaOrUnknown, PlanningProblemDetails.InvalidUseOfSet)
+                problemHandler.handleProblem(
+                    Problem(dmlOp.metas.sourceLocationMetaOrUnknown, PlanningProblemDetails.UnimplementedFeature("SET"))
                 )
+                INVALID_STATEMENT
             }
         }
+    }
+
+    override fun transformStatementDdl(node: PartiqlAst.Statement.Ddl): PartiqlLogical.Statement {
+        // It is an open question whether the planner will support DDL statements directly or if they must be handled by
+        // some other construct.  For now, we just submit an error with problem details indicating these statements
+        // are not implemented.
+        problemHandler.handleProblem(
+            Problem(
+                node.metas.sourceLocationMetaOrUnknown,
+                PlanningProblemDetails.UnimplementedFeature(
+                    when (node.op) {
+                        is PartiqlAst.DdlOp.CreateIndex -> "CREATE INDEX"
+                        is PartiqlAst.DdlOp.CreateTable -> "CREATE TABLE"
+                        is PartiqlAst.DdlOp.DropIndex -> "DROP INDEX"
+                        is PartiqlAst.DdlOp.DropTable -> "DROP TABLE"
+                    }
+                )
+            )
+        )
+        return INVALID_STATEMENT
     }
 
     override fun transformStatementDdl(node: PartiqlAst.Statement.Ddl): PartiqlLogical.Statement = abortQueryPlanning(
@@ -232,13 +271,22 @@ private object AstToLogicalVisitorTransform : PartiqlAstToPartiqlLogicalVisitorT
         }
 }
 
-private object FromSourceToBexpr : PartiqlAst.FromSource.Converter<PartiqlLogical.Bexpr> {
+private fun PartiqlAst.FromSource.toBexpr(
+    toLogicalTransform: AstToLogicalVisitorTransform,
+    problemHandler: ProblemHandler
+) =
+    FromSourceToBexpr(toLogicalTransform, problemHandler).convert(this)
+
+private class FromSourceToBexpr(
+    val toLogicalTransform: AstToLogicalVisitorTransform,
+    val problemHandler: ProblemHandler
+) : PartiqlAst.FromSource.Converter<PartiqlLogical.Bexpr> {
 
     override fun convertScan(node: PartiqlAst.FromSource.Scan): PartiqlLogical.Bexpr {
         val asAlias = node.asAlias ?: errAstNotNormalized("Expected as alias to be non-null")
         return PartiqlLogical.build {
             scan(
-                AstToLogicalVisitorTransform.transformExpr(node.expr),
+                toLogicalTransform.transformExpr(node.expr),
                 varDecl_(asAlias, asAlias.metas),
                 node.atAlias?.let { varDecl_(it, it.metas) },
                 node.byAlias?.let { varDecl_(it, it.metas) },
@@ -247,21 +295,32 @@ private object FromSourceToBexpr : PartiqlAst.FromSource.Converter<PartiqlLogica
         }
     }
 
-    override fun convertUnpivot(node: PartiqlAst.FromSource.Unpivot): PartiqlLogical.Bexpr {
-        abortUnimplementedFeature(node, "UNPIVOT")
-    }
+    override fun convertUnpivot(node: PartiqlAst.FromSource.Unpivot): PartiqlLogical.Bexpr =
+        INVALID_BEXPR.also { problemHandler.handleUnimplementedFeature(node, "UNPIVOT") }
 
     override fun convertJoin(node: PartiqlAst.FromSource.Join): PartiqlLogical.Bexpr =
         PartiqlLogical.build {
             join(
-                joinType = AstToLogicalVisitorTransform.transformJoinType(node.type),
+                joinType = toLogicalTransform.transformJoinType(node.type),
                 left = convert(node.left),
                 right = convert(node.right),
-                predicate = node.predicate?.let { AstToLogicalVisitorTransform.transformExpr(it) },
+                predicate = node.predicate?.let { toLogicalTransform.transformExpr(it) },
                 node.metas
             )
         }
 
     override fun convertGraphMatch(node: PartiqlAst.FromSource.GraphMatch): PartiqlLogical.Bexpr =
-        abortUnimplementedFeature(node, "MATCH")
+        INVALID_BEXPR.also { problemHandler.handleUnimplementedFeature(node, "MATCH") }
+}
+
+private val INVALID_STATEMENT = PartiqlLogical.build {
+    query(lit(ionSymbol("this is a placeholder for an invalid statement - do not run")))
+}
+
+private val INVALID_BEXPR = PartiqlLogical.build {
+    scan(lit(ionSymbol("this is a placeholder for an invalid relation - do not run")), varDecl("invalid"))
+}
+
+private val INVALID_EXPR = PartiqlLogical.build {
+    lit(ionSymbol("this is a placeholder for an invalid expression - do not run"))
 }