+++ lang/test/org/partiql/lang/eval/EvaluatorTests.kt
@@ -107,7 +107,8 @@ class EvaluatorTests {
             "undefinedUnqualifiedVariableInSelectWithUndefinedVariableBehaviorMissing",
 
             // we are currently not plumbed to be able to return the original letter casing of global variables.
-            // (there are other tests in LogicalToLogicalResolvedVisitorTransform which cover this case)
+            // (there are other tests in LogicalToLogicalResolvedVisitorTransform which cover this case for the
+            // PlannerPipeline)
             "identifierCaseMismatch"
         )